backend/requirements.txt
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
Flask-CORS==3.0.10
Werkzeug==2.0.3
gunicorn==20.1.0

